BC-317: Refactored view model tests.
