+++ src/main/resources/messages.properties
@@ -181,6 +181,7 @@ claimSummary.assessmentButton=Add assessment outcome
 claimSummary.update.assessmentButton=Update assessment outcome
 claimSummary.totalAssessmentText=Total authorised after assessment incl VAT:{0}
 claimSummary.lastAssessmentText=Last edited by {0} on {1} at {2}<br/>{3}.
+claimSummary.lastAssessmentText.noUser=Last edited on {0} at {1}<br/>{2}.
 claimSummary.alert.h2=This claim has been assessed
 claimSummary.alert.ariaLabel=information: This claim has been assessed
