+++ src/test/java/uk/gov/justice/laa/amend/claim/viewmodels/ClaimDetailsViewTest.java
@@ -1,0 +1,221 @@
+package uk.gov.justice.laa.amend.claim.viewmodels;
+
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Nested;
+import org.junit.jupiter.api.Test;
+import uk.gov.justice.laa.amend.claim.models.AssessmentInfo;
+import uk.gov.justice.laa.amend.claim.models.ClaimDetails;
+import uk.gov.justice.laa.amend.claim.models.ClaimField;
+import uk.gov.justice.laa.amend.claim.models.ClaimFieldStatus;
+import uk.gov.justice.laa.amend.claim.models.MicrosoftApiUser;
+import uk.gov.justice.laa.amend.claim.models.OutcomeType;
+
+import java.time.LocalDateTime;
+import java.time.OffsetDateTime;
+import java.time.YearMonth;
+import java.time.ZoneOffset;
+import java.util.List;
+
+public abstract class ClaimDetailsViewTest<C extends ClaimDetails, V extends ClaimDetailsView<C>> {
+
+    protected abstract C createClaim();
+    protected abstract V createView(C claim);
+
+    public static ClaimField createClaimField(String key, ClaimFieldStatus status) {
+        ClaimField claimField = new ClaimField();
+        claimField.setKey(key);
+        claimField.setStatus(status);
+        return claimField;
+    }
+
+    @Nested
+    class GetAccountNumberTests {
+        @Test
+        void getAccountNumberPicksOutFirstPartOfScheduleReference() {
+            C claim = createClaim();
+            claim.setScheduleReference("0U733A/2018/02");
+            V viewModel = createView(claim);
+            Assertions.assertEquals("0U733A", viewModel.getAccountNumber());
+        }
+
+        @Test
+        void getAccountNumberReturnsScheduleReferenceIfUnexpectedFormat() {
+            C claim = createClaim();
+            claim.setScheduleReference("0U733A201802");
+            V viewModel = createView(claim);
+            Assertions.assertEquals("0U733A201802", viewModel.getAccountNumber());
+        }
+    }
+
+    @Nested
+    class GetSubmissionPeriodForDisplayTests {
+        @Test
+        void getSubmissionPeriodForDisplayHandlesNull() {
+            C claim = createClaim();
+            V viewModel = createView(claim);
+            Assertions.assertNull(viewModel.getSubmissionPeriodForDisplay());
+        }
+
+        @Test
+        void getSubmissionPeriodForDisplayFormatsDate() {
+            C claim = createClaim();
+            claim.setSubmissionPeriod(YearMonth.of(2020, 1));
+            V viewModel = createView(claim);
+            Assertions.assertEquals("Jan 2020", viewModel.getSubmissionPeriodForDisplay());
+        }
+    }
+
+    @Nested
+    class GetSubmissionPeriodForSortingTests {
+        @Test
+        void getSubmissionPeriodForSortingHandlesNull() {
+            C claim = createClaim();
+            V viewModel = createView(claim);
+            Assertions.assertEquals(0, viewModel.getSubmissionPeriodForSorting());
+        }
+
+        @Test
+        void getSubmissionPeriodForSortingGetsEpochValueOfDate() {
+            C claim = createClaim();
+            claim.setSubmissionPeriod(YearMonth.of(2020, 1));
+            V viewModel = createView(claim);
+            Assertions.assertEquals(18262, viewModel.getSubmissionPeriodForSorting());
+        }
+    }
+
+    @Nested
+    class GetAssessedTotalsTests {
+        @Test
+        void getAssessedTotalsHandlesNullFields() {
+            C claim = createClaim();
+            V viewModel = createView(claim);
+            Assertions.assertEquals(List.of(), viewModel.getAssessedTotals());
+        }
+
+        @Test
+        void getAssessedTotalsHandlesValidFields() {
+            C claim = createClaim();
+            claim.setAssessedTotalVat(createClaimField("assessedTotalVat", ClaimFieldStatus.MODIFIABLE));
+            claim.setAssessedTotalInclVat(createClaimField("assessedTotalInclVat", ClaimFieldStatus.MODIFIABLE));
+            V viewModel = createView(claim);
+
+            List<ClaimField> result = viewModel.getAssessedTotals();
+            Assertions.assertEquals(claim.getAssessedTotalVat(), result.get(0));
+            Assertions.assertEquals(claim.getAssessedTotalInclVat(), result.get(1));
+        }
+
+        @Test
+        void getAssessedTotalsHandlesDoNotDisplayFields() {
+            C claim = createClaim();
+            claim.setAssessedTotalVat(createClaimField("assessedTotalVat", ClaimFieldStatus.DO_NOT_DISPLAY));
+            claim.setAssessedTotalInclVat(createClaimField("assessedTotalInclVat", ClaimFieldStatus.DO_NOT_DISPLAY));
+            V viewModel = createView(claim);
+
+            Assertions.assertEquals(List.of(), viewModel.getAssessedTotals());
+        }
+    }
+
+    @Nested
+    class GetAllowedTotalsTests {
+        @Test
+        void getAllowedTotalsHandlesNull() {
+            C claim = createClaim();
+            V viewModel = createView(claim);
+
+            Assertions.assertEquals(List.of(), viewModel.getAllowedTotals());
+        }
+
+        @Test
+        void getAllowedTotalsHandlesValid() {
+            C claim = createClaim();
+            claim.setAllowedTotalVat(createClaimField("allowedTotalVat", ClaimFieldStatus.MODIFIABLE));
+            claim.setAllowedTotalInclVat(createClaimField("allowedTotalInclVat", ClaimFieldStatus.MODIFIABLE));
+            V viewModel = createView(claim);
+
+            List<ClaimField> result = viewModel.getAllowedTotals();
+
+            Assertions.assertEquals(claim.getAllowedTotalVat(), result.get(0));
+            Assertions.assertEquals(claim.getAllowedTotalInclVat(), result.get(1));
+        }
+    }
+
+    @Nested
+    class GetClientNameTests {
+        @Test
+        void getClientNameHandlesNullForenameAndSurname() {
+            C claim = createClaim();
+            V viewModel = createView(claim);
+            Assertions.assertNull(viewModel.getClientName());
+        }
+
+        @Test
+        void getClientNameHandlesNullSurname() {
+            C claim = createClaim();
+            claim.setClientForename("John");
+            V viewModel = createView(claim);
+            Assertions.assertEquals("John", viewModel.getClientName());
+        }
+
+        @Test
+        void getClientNameHandlesNullForename() {
+            C claim = createClaim();
+            claim.setClientSurname("Doe");
+            V viewModel = createView(claim);
+            Assertions.assertEquals("Doe", viewModel.getClientName());
+        }
+
+        @Test
+        void getClientNameHandlesFullName() {
+            C claim = createClaim();
+            claim.setClientForename("John");
+            claim.setClientSurname("Doe");
+            V viewModel = createView(claim);
+            Assertions.assertEquals("John Doe", viewModel.getClientName());
+        }
+    }
+
+    @Nested
+    class LastEditedByTests {
+        @Test
+        void displayLastEditedTextWhenEverythingNonNull() {
+            C claim = createClaim();
+            AssessmentInfo assessmentInfo = new AssessmentInfo();
+            LocalDateTime localDateTime = LocalDateTime.of(2025, 12, 18, 16, 11, 27);
+            assessmentInfo.setLastAssessmentDate(OffsetDateTime.of(localDateTime, ZoneOffset.UTC));
+            assessmentInfo.setLastAssessmentOutcome(OutcomeType.NILLED);
+            claim.setLastAssessment(assessmentInfo);
+            V viewModel = createView(claim);
+            MicrosoftApiUser user = new MicrosoftApiUser("id", "Joe Bloggs");
+
+            ThymeleafMessage result = viewModel.lastEditedBy(user);
+
+            Assertions.assertEquals("claimSummary.lastAssessmentText", result.getKey());
+            Assertions.assertEquals("Joe Bloggs", result.getParams()[0]);
+            Assertions.assertEquals("18 December 2025", result.getParams()[1]);
+            Assertions.assertEquals("16:11:27", result.getParams()[2]);
+            ThymeleafMessage param = (ThymeleafMessage) result.getParams()[3];
+            Assertions.assertEquals("outcome.nilled", param.getKey());
+            Assertions.assertEquals(0, param.getParams().length);
+        }
+
+        @Test
+        void displayLastEditedTextWhenUserIsNull() {
+            C claim = createClaim();
+            AssessmentInfo assessmentInfo = new AssessmentInfo();
+            LocalDateTime localDateTime = LocalDateTime.of(2025, 12, 18, 16, 11, 27);
+            assessmentInfo.setLastAssessmentDate(OffsetDateTime.of(localDateTime, ZoneOffset.UTC));
+            assessmentInfo.setLastAssessmentOutcome(OutcomeType.NILLED);
+            claim.setLastAssessment(assessmentInfo);
+            V viewModel = createView(claim);
+
+            ThymeleafMessage result = viewModel.lastEditedBy(null);
+
+            Assertions.assertEquals("claimSummary.lastAssessmentText.noUser", result.getKey());
+            Assertions.assertEquals("18 December 2025", result.getParams()[0]);
+            Assertions.assertEquals("16:11:27", result.getParams()[1]);
+            ThymeleafMessage param = (ThymeleafMessage) result.getParams()[2];
+            Assertions.assertEquals("outcome.nilled", param.getKey());
+            Assertions.assertEquals(0, param.getParams().length);
+        }
+    }
+}
